Remove unnecessary metrics. Move log_request to log. It also updates main for it.
